feat HttpRequestFactory Test 추가
