+++ src/test/resources/Http_POST.txt
@@ -1,4 +1,4 @@
-POST /user/create?id=1 HTTP/1.1
+POST /user/create HTTP/1.1
 Host: localhost:8080
 Connection: keep-alive
 Content-Length: 46
